Enable token authentication for API calls through devise
Devise doesn't allow passing the token through a HTTP header by default.
In previous versions of Cartoque this was simply done with a
before_filter but it seems this technique doesn't work anymore, probably
because warden has already parsed parameters when the filter is
triggered. Hence the DeviseHeaderToken rack middleware which adds an
artificial parameter to the request if the header is present.